move resonator_spectroscopy regression test to its own module
